+++ packages/components/molecules/f-mega-modal/test/accessibility/axe-accessibility.spec.js
@@ -1,17 +1,10 @@
-const { getAxeResults } = require('../../../../../../test/utils/axe-helper');
-
-const MegaModal = require('../../test-utils/component-objects/f-mega-modal.component');
-
-const megaModal = new MegaModal();
+import MegaModal from '../../test-utils/component-objects/f-mega-modal.component';
 
 describe('Accessibility tests', () => {
-    beforeEach(() => {
-        megaModal.load();
-    });
-
-    it('a11y - should test f-mega-modal component WCAG compliance', () => {
+    it('a11y - should test f-mega-modal component WCAG compliance', async () => {
         // Act
-        const axeResults = getAxeResults('f-mega-modal');
+        await MegaModal.load();
+        const axeResults = await MegaModal.getAxeResults('f-mega-modal');
 
         // Assert
         expect(axeResults.violations.length).toBe(0);