+++ packages/components/molecules/f-navigation-links/test/accessibility/axe-accessibility.spec.js
@@ -1,19 +1,10 @@
-import { getAxeResults } from '../../../../../../test/utils/axe-helper'; // eslint-disable-line import/no-relative-packages
-
-const NavigationLinks = require('../../test-utils/component-objects/f-navigation-links.component');
-
-let navigationLinks;
+import NavigationLinks from '../../test-utils/component-objects/f-navigation-links.component';
 
 describe('Accessibility tests', () => {
-    beforeEach(async () => {
-        navigationLinks = new NavigationLinks();
-
-        await navigationLinks.load();
-    });
-
-    it('a11y - should test f-navigation-links component WCAG compliance', () => {
-        // Act
-        const axeResults = getAxeResults('f-navigation-links');
+    it('a11y - should test f-navigation-links component WCAG compliance', async () => {
+        // Arrange
+        await NavigationLinks.load();
+        const axeResults = await NavigationLinks.getAxeResults('f-navigation-links');
 
         // Assert
         expect(axeResults.violations.length).toBe(0);