+++ packages/components/molecules/f-promotions-showcase/test/accessibility/axe-accessibility.spec.js
@@ -1,16 +1,10 @@
-import { getAxeResults } from '../../../../../../test/utils/axe-helper'; // eslint-disable-line import/no-relative-packages
-
-const PromotionsShowcase = require('../../test-utils/component-objects/f-promotions-showcase.component');
-
-const promotionsShowcase = new PromotionsShowcase();
+import PromotionsShowcase from '../../test-utils/component-objects/f-promotions-showcase.component';
 
 describe('Accessibility tests', () => {
-    beforeEach(async () => {
-        await promotionsShowcase.load();
-    });
-    it('a11y - should test f-promotions-showcase component WCAG compliance', () => {
+    it('a11y - should test f-promotions-showcase component WCAG compliance', async () => {
         // Act
-        const axeResults = getAxeResults('f-promotions-showcase');
+        await PromotionsShowcase.load();
+        const axeResults = await PromotionsShowcase.getAxeResults('f-promotions-showcase');
 
         // Assert
         expect(axeResults.violations.length).toBe(0);