+++ packages/storybook/vue.config.js
@@ -48,7 +48,9 @@ module.exports = {
                         'templates',
                         'f-account-info',
                         'f-checkout',
-                        'f-contact-preferences'
+                        'f-contact-preferences',
+                        'f-loyalty',
+                        'f-offers'
                     ];
                     const pathContainsUpdatedComponentOrType = updateComponentsAndTypes.some(a => absPathCommonScss.includes(a));