+++ packages/components/pages/f-offers/CHANGELOG.md
@@ -4,11 +4,20 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 
-v1.12.0-beta.3
-------------------------------
-*July 8, 2021*
+v2.0.0-beta.2
+-----------------------------
+*July 14, 2022*
 
 ### Changed
+- Move fozzie components to peer dependencies.
+
+
+v2.0.0-beta.1
+-----------------------------
+*July 12, 2022*
+
+### Changed
+- **Breaking changes:** Update to `@use` and `@forward` SASS syntax
 - Refactor WebDriverIO tests to use async in order to support Node 16 using `codemod` utility.